+++ src/main/java/Figures/Line/Line2d.java
@@ -33,7 +33,7 @@ public class Line2d implements Drawable {
 
 
     @Override
-    public void setColor(float red, float green, float blue) {
+    public void setColor(int red, int green, int blue) {
         color.setRed(red);
         color.setGreen(green);
         color.setBlue(blue);
