+++ src/main/java/Interfaces/Drawable.java
@@ -3,7 +3,7 @@ package Interfaces;
 public interface Drawable {
     public void draw();
 
-    default public void setColor(float red, float green, float blue) {
+    default public void setColor(int red, int green, int blue) {
     }
 
     public Drawable setScale(float scale);
